Refactored (moved) db models for funds and proposals testing components to models modules
