+++ src/v0/endpoints/funds/handlers.rs
@@ -14,43 +14,12 @@ pub async fn get_fund(context: SharedContext) -> Result<impl Reply, Rejection> {
 #[cfg(test)]
 pub mod test {
     use super::*;
-    use crate::db::{models::funds::Fund, schema::funds, testing as db_testing, DBConnectionPool};
+    use crate::db::{
+        models::funds::{test as funds_testing, Fund},
+        testing as db_testing,
+    };
     use crate::v0::context::test::new_in_memmory_db_test_shared_context;
-
-    use chrono::Utc;
-    use diesel::{ExpressionMethods, RunQueryDsl};
     use warp::Filter;
-
-    pub fn get_test_fund() -> Fund {
-        Fund {
-            id: 1,
-            fund_name: "hey oh let's go".to_string(),
-            fund_goal: "test this endpoint".to_string(),
-            voting_power_info: ">9000".to_string(),
-            rewards_info: "not much".to_string(),
-            fund_start_time: Utc::now().timestamp(),
-            fund_end_time: Utc::now().timestamp(),
-            next_fund_start_time: Utc::now().timestamp(),
-            chain_vote_plans: vec![],
-        }
-    }
-
-    pub fn populate_db_with_fund(fund: &Fund, pool: &DBConnectionPool) {
-        let connection = pool.get().unwrap();
-        let values = (
-            funds::fund_name.eq(fund.fund_name.clone()),
-            funds::fund_goal.eq(fund.fund_goal.clone()),
-            funds::voting_power_info.eq(fund.voting_power_info.clone()),
-            funds::rewards_info.eq(fund.rewards_info.clone()),
-            funds::fund_start_time.eq(fund.fund_start_time),
-            funds::fund_end_time.eq(fund.fund_end_time),
-            funds::next_fund_start_time.eq(fund.next_fund_start_time),
-        );
-        diesel::insert_into(funds::table)
-            .values(values)
-            .execute(&connection)
-            .unwrap();
-    }
 
     #[tokio::test]
     async fn get_fund_handler() {
@@ -62,8 +31,8 @@ pub mod test {
         // initialize db
         let pool = &shared_context.read().await.db_connection_pool;
         db_testing::initialize_db_with_migration(&pool);
-        let fund: Fund = get_test_fund();
-        populate_db_with_fund(&fund, &pool);
+        let fund: Fund = funds_testing::get_test_fund();
+        funds_testing::populate_db_with_fund(&fund, &pool);
 
         // build filter
         let filter = warp::any()
@@ -88,8 +57,8 @@ pub mod test {
         // initialize db
         let pool = &shared_context.read().await.db_connection_pool;
         db_testing::initialize_db_with_migration(&pool);
-        let fund: Fund = get_test_fund();
-        populate_db_with_fund(&fund, &pool);
+        let fund: Fund = funds_testing::get_test_fund();
+        funds_testing::populate_db_with_fund(&fund, &pool);
 
         // build filter
         let filter = warp::path!(i32)
